--- a/.changeset/chilly-items-help.md
+++ b/.changeset/chilly-items-help.md
@@ -1,0 +1,5 @@
+---
+"astro": patch
+---
+
+Improves the error message when failed to render MDX components