--- a/packages/astro/src/jsx/server.ts
+++ b/packages/astro/src/jsx/server.ts
@@ -4,6 +4,8 @@ import { renderJSX } from '../runtime/server/jsx.js';
 
 const slotName = (str: string) => str.trim().replace(/[-_]([a-z])/g, (_, w) => w.toUpperCase());
 
+// NOTE: In practice, MDX components are always tagged with `__astro_tag_component__`, so the right renderer
+// is used directly, and this check is not often used to return true.
 export async function check(
 	Component: any,
 	props: any,
@@ -19,18 +21,7 @@ export async function check(
 		const result = await Component({ ...props, ...slots, children });
 		return result[AstroJSX];
 	} catch (e) {
-		const error = e as Error;
-		// if the exception is from an mdx component
-		// throw an error
-		if (Component[Symbol.for('mdx-component')]) {
-			throw new AstroError({
-				message: error.message,
-				title: error.name,
-				hint: `This issue often occurs when your MDX component encounters runtime errors.`,
-				name: error.name,
-				stack: error.stack,
-			});
-		}
+		throwEnhancedErrorIfMdxComponent(e as Error, Component);
 	}
 	return false;
 }
@@ -48,8 +39,27 @@ export async function renderToStaticMarkup(
 	}
 
 	const { result } = this;
-	const html = await renderJSX(result, jsx(Component, { ...props, ...slots, children }));
-	return { html };
+	try {
+		const html = await renderJSX(result, jsx(Component, { ...props, ...slots, children }));
+		return { html };
+	} catch (e) {
+		throwEnhancedErrorIfMdxComponent(e as Error, Component);
+		throw e;
+	}
+}
+
+function throwEnhancedErrorIfMdxComponent(error: Error, Component: any) {
+	// if the exception is from an mdx component
+	// throw an error
+	if (Component[Symbol.for('mdx-component')]) {
+		throw new AstroError({
+			message: error.message,
+			title: error.name,
+			hint: `This issue often occurs when your MDX component encounters runtime errors.`,
+			name: error.name,
+			stack: error.stack,
+		});
+	}
 }
 
 export default {