--- a/packages/astro/src/jsx/transform-options.ts
+++ b/packages/astro/src/jsx/transform-options.ts
@@ -1,5 +1,8 @@
 import type { JSXTransformConfig } from '../@types/astro.js';
 
+/**
+ * @deprecated This function is no longer used. Remove in Astro 5.0
+ */
 export async function jsxTransformOptions(): Promise<JSXTransformConfig> {
 	// @ts-expect-error types not found
 	const plugin = await import('@babel/plugin-transform-react-jsx');