… .changeset/blue-geese-visit.md
… .changeset/fresh-masks-agree.md
… .changeset/friendly-plants-leave.md
… .changeset/large-glasses-jam.md
… .changeset/slimy-cobras-end.md
… .changeset/small-oranges-report.md
… .changeset/smart-rats-mate.md
… .changeset/sweet-goats-own.md
… .changeset/tame-avocados-relax.md
… .changeset/violet-snails-call.md
… .changeset/young-chicken-exercise.md
… packages/astro/package.json
… packages/astro/test/units/dev/collections-renderentry.test.js
… packages/integrations/mdx/package.json
… packages/integrations/mdx/src/README.md
… packages/integrations/mdx/src/remark-images-to-component.ts
… packages/integrations/mdx/src/vite-plugin-mdx-postprocess.ts
… packages/integrations/mdx/src/vite-plugin-mdx.ts
… packages/integrations/mdx/test/fixtures/mdx-optimize/astro.config.mjs
… packages/integrations/mdx/test/mdx-plugins.test.js
… packages/integrations/mdx/test/units/rehype-optimize-static.test.js
… pnpm-lock.yaml