## Changes

This PR is a feature branch merging all the existing work for the new MDX integration v3 major. It also adds code to `astro` and requires a minor bump of it. The MDX integration will depend on the latest Astro version on publish, which is v4.8.0.

All merged PRs (reviewed, chronological):
1. https://github.com/withastro/astro/pull/10697
2. https://github.com/withastro/astro/pull/10866
3. https://github.com/withastro/astro/pull/10864
4. https://github.com/withastro/astro/pull/10869
5. https://github.com/withastro/astro/pull/10885
6. https://github.com/withastro/astro/pull/10884
7. https://github.com/withastro/astro/pull/10877
8. https://github.com/withastro/astro/pull/10887
9. https://github.com/withastro/astro/pull/10934

Other commits I added (unreviewed):
1. https://github.com/withastro/astro/commit/947b133752d920c3ab3997587d4ca2e37c2f180b: Improve types in `rehype-optimize-static.ts`. No logic changed so I figured it didn't need a review.
2. https://github.com/withastro/astro/commit/1d0c6bec306e8a83a750a3f7a6038d6ffeeaf7c3: Rename `ignoreComponentNames` to `ignoreElementNames`. I think it better reflected the intent. This was also renamed in PR no6 above.

Performance:

<table>
<tr>
 <td>
 <td>Rollup
 <td>Render
 <td>Total
<tr>
 <td>Before
 <td>101.5s
 <td>38s
 <td>149.5s
<tr>
 <td>After
 <td>85.5s
 <td>37.5s
 <td>123s
<tr>
 <td>Improvement
 <td>16% faster
 <td>1% faster
 <td><b>18% faster</b>
</table>

Thoughts: I expected more to be honest but I'll take what I can 😅 

<details>
<summary>Raw data</summary>

```
before 1: build 99s. render 49s
before 2: build 104s. render 33s
before 3: build 104s. render 43s

after 1: build 86s. render 48s
after 2: build 88s. render 32s
after 3: build 85s. render 43s
```

I took the two fastest timing for each build/render phase and average them. As the render phase are dependent of network and fluctuate a bit.

</details>

## Testing

Existing tests should pass. New tests added from PRs should also pass here.

## Docs

Updated MDX integration page: https://github.com/withastro/docs/pull/8151